Added observation tables, implemented Printing Errors. Fixed bugs in validator.js and layout.js, rewrote procedure and removed unused images.
