+++ experiment/images/nor_latch.png
