+++ experiment/pretest.json
@@ -58,7 +58,7 @@
             "difficulty": "Intermediate"
         },
         {
-            "question": "5. THe flip-flop is a _________ device ?",
+            "question": "5. THe flip-flop is a/an _________ device ?",
             "answers": {
                 "a": "unstable",
                 "b": "bistable",
