+++ experiment/simulation/demo-TFF.html
@@ -116,7 +116,7 @@
         </div>
     </div>
     <script src="https://cdn.jsdelivr.net/gh/virtual-labs/virtual-style@0.0.8-b/js/script.js"></script>
-    <script type="module" src="js/anim.js"></script>
+    <script type="module" src="js/animation-TFF.js"></script>
 
 </body>
