--- experiment/simulation/js/anim.js
@@ -1,400 +1,0 @@
-import { setCoordinates,fillInputDots,fillColor,objectDisappear,objectAppear,setter, calculateNot, calculateAnd} from "./animation-utility.js";
-
-'use strict'
-
-window.appendToggle = appendToggle;
-window.appendClock = appendClock;
-window.simulationStatus = simulationStatus;
-window.restartCircuit = restartCircuit;
-window.setSpeed=setSpeed;
-
-
-
-// Dimensions of working area
-const circuitBoard = document.getElementById("circuit-board");
-const sidePanels = document.getElementsByClassName("v-datalist-container");
-// Distance of working area from top
-const circuitBoardTop = circuitBoard.offsetTop;
-// Full height of window
-const windowHeight = window.innerHeight;
-const width = window.innerWidth;
-
-const svg = document.querySelector(".svg");
-const svgns = "http://www.w3.org/2000/svg";
-
-const EMPTY="";
-// stroing the necessary div elements in const
-const status = document.getElementById("play-or-pause");
-const observ = document.getElementById("observations");
-const speed = document.getElementById("speed");
-
-// global varaibles declared here
-const objects = [
-    document.getElementById("toggle"),
-    document.getElementById("clock"),
-    document.getElementById("output-q"),
-    document.getElementById("output-qbar")
-];
-const textInput = [
-    document.createElementNS(svgns, "text"),
-    document.createElementNS(svgns, "text")
-];
-const textOutput = [
-    document.createElementNS(svgns, "text"),
-    document.createElementNS(svgns, "text")
-];
-const dots = [
-    document.createElementNS(svgns, "circle"),
-    document.createElementNS(svgns, "circle"),
-    document.createElementNS(svgns, "circle")
-];
-// First 2 dots emerge from Data
-// Third dot is from Clock
-
-
-
-// decide help to decide the speed
-let decide = false;
-// circuitStarted is initialised to 0 which depicts that demo hasn't started whereas circuitStarted 1 depicts that the demo has started.
-let circuitStarted = false;
-
-
-// function to take care of width
-function demoWidth() {
-    if (width < 1024) {
-        circuitBoard.style.height = "600px";
-    } else {
-        circuitBoard.style.height = `${windowHeight - circuitBoardTop - 20}px`;
-    }
-    sidePanels[0].style.height = circuitBoard.style.height;
-}
-
-// function to initialise the input text i.e. either 0/1 that gets displayed after user click on them
-function textIOInit() {
-    for( const text of textInput){
-        text.textContent = 2;
-    }
-}
-
-
-// function to mark the output coordinates
-function outputCoordinates() {
-    setCoordinates(896,424,textOutput[0]);
-    svg.append(textOutput[0]);
-    setCoordinates(896,604,textOutput[1]);
-    svg.append(textOutput[1]);
-}
-
-// function to mark the input dots
-function inputDots() {
-    for(const dot of dots){
-        fillInputDots(dot,20,550,15,"#FF0000");
-        svg.append(dot);
-    }
-}
-
-// function to disappear the input dots
-function inputDotDisappear() {
-    for(const dot of dots){
-        objectDisappear(dot);
-    }
-}
-
-// function to appear the input dots
-function inputDotVisible() {
-    for(const dot of dots){
-        objectAppear(dot);
-    }
-}
-// function to disappear the output text
-function outputDisappear() {
-    for(const text of textOutput){
-        objectDisappear(text);
-    }
-}
-// function to appear the output text
-function outputVisible() {
-    for(const text of textOutput){
-        objectAppear(text);
-    }
-}
-// function to diappear the input text
-function inputTextDisappear() {
-    for(const text of textInput){
-        objectDisappear(text);
-    }
-}
-
-function clearObservation() {
-    observ.innerHTML = EMPTY;
-}
-function allDisappear() {
-    inputDotDisappear();
-    outputDisappear();
-    inputTextDisappear();
-    for(const object of objects){
-        fillColor(object,"#008000");
-    }
-}
-
-function appendToggle() {
-    if (textInput[0].textContent !== "0" && timeline.progress() === 0) {
-        changeto0(16,124,0,0);
-        observ.innerHTML = "Data bit is equal to 0";
-    }
-    else if (textInput[0].textContent !== "1" && timeline.progress() === 0) {
-        changeto1(16,124,0,0);
-        observ.innerHTML = "Data bit is equal to 1";
-    }
-    setter(textInput[0].textContent,dots[0]);
-    setter(textInput[0].textContent,dots[1]);
-}
-function appendClock() {
-    if (textInput[1].textContent !== "0" && timeline.progress() === 0) {
-        changeto0(16,514,1,1);
-    }
-    else if (textInput[1].textContent !== "1" && timeline.progress() === 0) {
-        changeto1(16,514,1,1);
-        observ.innerHTML = "Clock is turned ON";
-    }
-    setter(textInput[1].textContent,dots[2]);
-}
-
-function changeto1(coordinateX,coordinateY,object,textObject) {
-    textInput[textObject].textContent = 1;
-    svg.appendChild(textInput[textObject]);
-    setCoordinates(coordinateX,coordinateY,textInput[textObject]);
-    fillColor(objects[object],"#29e");
-    clearObservation();
-    objectAppear(textInput[textObject]);
-}
-
-function changeto0(coordinateX,coordinateY,object,textObject) {
-    textInput[textObject].textContent = 0;
-    svg.appendChild(textInput[textObject]);
-    setCoordinates(coordinateX,coordinateY,textInput[textObject]);
-    fillColor(objects[object],"#eeeb22");
-    clearObservation();
-    objectAppear(textInput[textObject]);
-}
-
-let q = "0";
-let qbar = "0";
-
-function stage1() {
-    if(textInput[0].textContent === "0"){
-        q = textOutput[0].textContent;
-        qbar = textOutput[1].textContent;
-    }
-    else if(textInput[0].textContent === "1"){
-        q = textOutput[1].textContent;
-        qbar = textOutput[0].textContent;
-    }
-    else
-    {
-        console.log("Error! Unreachable state");
-    }
-    setter(q,dots[0]);
-    setter(qbar,dots[1]);
-    objectDisappear(dots[2]);
-
-}
-
-function outputSetter(){
-    textOutput[0].textContent = q;
-    textOutput[1].textContent = qbar;
-    setter(textOutput[0].textContent,objects[2]);
-    setter(textOutput[1].textContent,objects[3]);
-    objectDisappear(dots[0]);
-    objectDisappear(dots[1]);
-}
-
-function display() {
-    observ.innerHTML = "Simulation has finished. Press Restart to start again"
-}
-
-function reboot() {
-    for(const text of textInput){
-        text.textContent = 2;
-    }
-}
-
-function setSpeed(speed) {
-    if (circuitStarted) {
-        timeline.timeScale(parseInt(speed));
-        observ.innerHTML = `${speed}x speed`;
-    }
-}
-
-function restartCircuit() {
-    if (circuitStarted) {
-        circuitStarted = false;
-    }
-    timeline.seek(0);
-    timeline.pause();
-    allDisappear();
-    reboot();
-    clearObservation();
-    decide = false;
-    status.innerHTML = "Start";
-    observ.innerHTML = "Successfully restored. ";
-    observ.innerHTML += "Initially the circuit is in reset state. i.e. Q = 0 and Q' = 1";
-    speed.selectedIndex = 0;
-}
-
-function simulationStatus() {
-    if (!decide) {
-        startCircuit();
-    }
-    else {
-        stopCircuit();
-    }
-}
-function stopCircuit() {
-    if (timeline.progress() !== 1) {
-        timeline.pause();
-        observ.innerHTML = "Simulation has been stopped.";
-        decide = false;
-        status.innerHTML = "Start";
-        speed.selectedIndex = 0;
-    }
-    else {
-        observ.innerHTML = "Please Restart the simulation";
-    }
-}
-function startCircuit() {
-    if(textInput[1].textContent==="0")
-    {
-        observ.innerHTML = "Please set the Clock as 1";
-        return;
-    }
-    textOutput[0].textContent = "0";
-    textOutput[1].textContent = "1";
-
-
-    for(const text of textInput){
-        if (text.textContent === "2") {
-            observ.innerHTML = "Please set the input values";
-            return;
-        }
-    }
-    if (timeline.progress() !== 1) {
-        if (!circuitStarted) {
-            circuitStarted = true;
-        }
-        timeline.play();
-        timeline.timeScale(1);
-        observ.innerHTML = "Simulation has started.";
-        decide = true;
-        status.innerHTML = "Pause";
-        speed.selectedIndex = 0;
-    }
-    else {
-        observ.innerHTML = "Please Restart the simulation";
-    }
-}
-
-
-
-// all the execution begin here
-let timeline = gsap.timeline({ repeat: 0, repeatDelay: 0 });
-gsap.registerPlugin(MotionPathPlugin);
-demoWidth();
-// calling all the functions that are going to initialise 
-textIOInit();
-outputCoordinates();
-inputDots();
-outputDisappear();
-
-timeline.add(inputDotVisible, 0);
-timeline.add(stage1,10);
-timeline.add(outputSetter,15);
-timeline.add(outputVisible,15);
-timeline.eventCallback("onComplete", outputVisible);
-timeline.eventCallback("onComplete", display);
-
-timeline.to(dots[0], {
-    motionPath: {
-        path: "#path1",
-        align: "#path1",
-        autoRotate: true,
-        alignOrigin: [0.5, 0.5]
-    },
-
-    duration: 10,
-    repeat: 0,
-    repeatDelay: 3,
-    yoyo: true,
-    ease: "none",
-    paused: false,
-
-}, 0);
-timeline.to(dots[1], {
-    motionPath: {
-        path: "#path2",
-        align: "#path2",
-        autoRotate: true,
-        alignOrigin: [0.5, 0.5]
-    },
-
-    duration: 10,
-    repeat: 0,
-    repeatDelay: 3,
-    yoyo: true,
-    ease: "none",
-    paused: false,
-
-}, 0);
-timeline.to(dots[2], {
-    motionPath: {
-        path: "#path3",
-        align: "#path3",
-        autoRotate: true,
-        alignOrigin: [0.5, 0.5]
-    },
-
-    duration: 10,
-    repeat: 0,
-    repeatDelay: 3,
-    yoyo: true,
-    ease: "none",
-    paused: false,
-
-}, 0);
-timeline.to(dots[0], {
-    motionPath: {
-        path: "#path4",
-        align: "#path4",
-        autoRotate: true,
-        alignOrigin: [0.5, 0.5]
-    },
-
-    duration: 5,
-    delay: 10,
-    repeat: 0,
-    repeatDelay: 3,
-    yoyo: true,
-    ease: "none",
-    paused: false,
-
-}, 0);
-timeline.to(dots[1], {
-    motionPath: {
-        path: "#path5",
-        align: "#path5",
-        autoRotate: true,
-        alignOrigin: [0.5, 0.5]
-    },
-    
-    duration: 5,
-    delay: 10,
-    repeat: 0,
-    repeatDelay: 3,
-    yoyo: true,
-    ease: "none",
-    paused: false,
-
-},0);
-observ.innerHTML = "Initially the circuit is in reset state. i.e. Q = 0 and Q' = 1";
-timeline.pause();
-inputDotDisappear();
