+++ experiment/simulation/js/gate.js
@@ -8,11 +8,6 @@ import { checkConnectionsRS, checkConnectionsJK, simulateFFJK, simulateFFRS, tes
 export let gates = {}; // Array of gates
 window.numComponents = 0;
 export function clearGates() {
-
-    for (let gateId in gates) {
-        delete gates[gateId];
-    }
-
     gates = {};
 }
 
@@ -233,30 +228,41 @@ function setInput(event) {
 
 window.setInput = setInput;
 
-export function checkConnections() {
-    let correctConnection = true;
-    for (let gateId in gates) {
-        const gate = gates[gateId];
-        if (gate.inputPoints.length != gate.inputs.length) {
-            correctConnection = false;
-        }
-        else if (!gate.isConnected && !gate.isOutput) {
-            correctConnection = false;
-        }
-    }
-    if (correctConnection) {
-        return true;
-    }
-    else {
-        alert("Connections are not correct");
-        return false;
+export function clearResult() {
+    const result = document.getElementById("result");
+    result.innerHTML = "";
+}
+
+export function printErrors(message,objectId) {
+    const result = document.getElementById('result');
+    result.innerHTML += message;
+    result.className = "failure-message";
+    if(objectId !== null)
+    {
+        objectId.classList.add("highlight")
+        setTimeout(function () {objectId.classList.remove("highlight")}, 5000);
     }
 }
 
+export function checkConnections() {
+    for (let gateId in gates) {
+        const gate = gates[gateId];
+        const id = document.getElementById(gate.id);
+        if (gate.inputPoints.length != gate.inputs.length) {
+            printErrors("Highlighted component not connected properly\n",id);
+            return false;
+        }
+        else if (!gate.isConnected && !gate.isOutput) {
+            printErrors("Highlighted component not connected properly\n",id);
+            return false;
+        }
+    }
+    return true;
+}
+
 export function simulate() {
-
-    window.simulate = 0;
-
+    clearResult();
+    window.simulationStatus = 0;
     if (!checkConnections()) {
         return false;
     }
@@ -282,13 +288,13 @@ export function simulate() {
         }
     }
     if (!circuitHasClock) {
-        simulate2();
+        simulateWithClock();
     }
 
     return true;
 }
 
-function simulate2() {
+function simulateWithClock() {
     // input bits
     for (let gateId in gates) {
         const gate = gates[gateId];
@@ -416,22 +422,25 @@ function simulate2() {
     }
 }
 
-window.sim = simulate;
-window.sim2 = simulate2;
+window.simulate = simulate;
+window.simClk = simulateWithClock;
 
 export function testSimulation(gates,flipFlops) {
     if (!checkConnections()) {
-        return;
+        document.getElementById("table-body").innerHTML = "";
+        return false;
     }
 
     if (window.currentTab === "task2") {
         if (!checkConnectionsRS()) {
-            return;
+            document.getElementById("table-body").innerHTML = "";
+            return false;
         }
     }
     else if (window.currentTab === "task4") {
         if (!checkConnectionsJK()) {
-            return;
+            document.getElementById("table-body").innerHTML = "";
+            return false;
         }
     }
 
@@ -552,11 +561,12 @@ export function testSimulation(gates,flipFlops) {
             }
         }
     }
+    return true;
 }
 
 // function to submit the desired circuit and get the final success or failure message
 export function submitCircuit() {
-
+    clearResult();
     document.getElementById("table-body").innerHTML = "";
     if (window.currentTab === "task1") {
         testRSFF("Input-1", "Input-0", "Clock-0", "Output-2", "Output-3");
