+++ experiment/simulation/practice.html
@@ -88,19 +88,33 @@
 			<div class="v-datalist-container components-list">
 				<div class="v-datalist-title">Observations</div>
 				<div class="v-datalist-row">
+					<div id="result"></div>
 					<div class="v-table-wrap">
 						<table class="table is-bordered is-fullwidth">
 							<thead id="table-head">
+								<tr>
+									<th colspan="3">Inputs</th>
+									<th colspan="2">Expected Values</th>
+									<th colspan="2">Observed Values</th>
+								</tr>
+								<tr>
+									<th>R</th>
+									<th>S</th>
+									<th>Clk</th>
+									<th>Q</th>
+									<th>Q'</th>
+									<th>Q</th>
+									<th>Q'</th>
+								</tr>
 							</thead>
 							<tbody id="table-body">
 							</tbody>
 						</table>
 					</div>
-					<div id="result"></div>
 				</div>
 				<div class="v-datalist-row">
-					<button class="v-button" id="simulate-button">Simulate</button>
-					<button class="v-button" onclick="submitCircuit()">Submit</button>
+					<button class="v-button" id="simulate-button" onclick="toggleSimulation()">Simulate</button>
+					<button class="v-button" id="submit-button" onclick="submitCircuit()">Submit</button>
 				</div>
 
 			</div>
@@ -115,12 +129,14 @@
 	<div class="modal">
 		<div class="modal-content">
 			<div class="container">
-				<span class="close-button">&times;</span>
+				<div class="columns is-mobile is-flex-direction-row-reverse">
+					<span class="close-button">&times;</span>
+				</div>
 				<div class="columns is-centered is-mobile">
 					<div class="column is-one-quarter">Freq(Hz): </div>
 					<div class="column">
 						<div class="control">
-							<input class="input" id="frequency-input" type="text" placeholder="" />
+							<input class="input is-primary" id="frequency-input" type="text" placeholder="" />
 						</div>
 					</div>
 				</div>
@@ -128,7 +144,7 @@
 					<div class="column is-one-quarter">Duty Cycle(%): </div>
 					<div class="column">
 						<div class="control">
-							<input class="input" id="dutycycle-input" type="text" placeholder="" />
+							<input class="input is-primary" id="dutycycle-input" type="text" placeholder="" />
 						</div>
 					</div>
 				</div>
