+++ experiment/theory.md
@@ -2,14 +2,14 @@
 
 **NOR Latch**
 
-<img src="images/NOR.png">
+<img src="images/nor_latch.png">
 
 2-d vectors
 It can be constructed from a pair of cross-coupled NOR logic gates. The stored bit is present on the output marked Q. While the S and R inputs are both low, feedback maintains the Q and Q outputs in a constant state, with Q the complement of Q. If S (Set) is pulsed high while R (Reset) is held low, then the Q output is forced high, and stays high when S returns to low; similarly, if R is pulsed high while S is held low, then the Q output is forced low, and stays low when R returns to low. 
 
 **NAND Latch**
 
-<img src="images/NAND.png">
+<img src="images/nand_latch.png">
 
 This is an alternate model of the simple SR latch built with NAND logic gates. Set and reset now become active low signals, denoted S and R respectively. Otherwise, operation is identical to that of the SR latch. Historically, SR-latches have been predominant despite the notational inconvenience of active-low inputs 
 
@@ -24,7 +24,7 @@ JK flip flop is mostly implemented as Master Slave flip flop. The information pr
 
 The following Digram depicts the change of JK flip flop Output with time pulse. It is negativve edge triggered. 
 
-<img src="images/JK_flipflop_timing.gif">
+<img src="images/jk_flipflop_timing.gif">
 
 
 **Truth Table**
